refactor(KonnectorAction): Split match function
