+++ src/ducks/transactions/actions/KonnectorAction/helpers.js
@@ -1,0 +1,3 @@
+export function getBrandsWithoutTrigger(brands) {
+  return brands.filter(brand => !brand.hasTrigger)
+}
